Don't add continuation when 2nd defendant is blank
There is case where the blank fields of the defendant model
where being concatenated and added to the PDF form, so that
the continuation sheet was being appended.

This change is a prevention of that. The reason for checking
only of title & full_name if they are blank, is because the
model's validations handle that already. The check is only in
'#as_json'.