Cleaner way to diff PDFs
Previously two defendants were populated into the PDF, which would
add a continuation sheet. Which in turn would be considered as a
creation of a new PDF document. PDFs contain a timestamp when they
were created, so when diff-ing two PDFs, they could never be the
same just for this reason.

The workarounds for it were dirty, so thats why I'm testing the
population of a PDF without a continuation. Then the continuation
is being tested if it has the correct number of pages.