Reverting back all the 'no's to 'No's
There was only one exception where the *No* value was turned into a
lowercase 'no'. The fix would've been to revert only that change.  But
since all the 'No's (uppercase) were turned into 'no's (lowercase), this
change reverts them back to the correct state. As all the 'Yes' options
are already capitalised in the views as well as in the PDF. The same
applies to the 'No' (options).