AD tightening
Changed the field in Zendesk that stored the phone number from a
strictly numeric field to a string field. The reason for that is so that
the validation of the input only happens once, in the Rails app, rather
than twice in the Rails app & Zendesk.

Also tightened the validation of the phone number as it was allowing
alphabetic characters.