Fetch and populate Court Address details
CoffeeScript for fetching the nearest court address via the Court Finder
API. The tests for this need to be revisited and some edge cases need to
be considered like:

When you look up a postcode and the result is not found.

The text displayed is still showing the link 'Choose to send this
claim to a different court'. Should that link be omitted completely
and the form for the court address shown fully for manual
population?