Fix 'vagrant provision' for docker build/run
The docker 'reprovision' is not clever enough to kill it's previous
incarnations, so manually do this in our docker 'setup step'.

This means 'vagrant provision' will rebuild and rerun the end-result
container, allowing us to iterate on the Dockerfile without destroy/up.

This takes the process down to 8m-ish, over 20m-ish.