Updated the main PDF form
The heading in the section 4 also needs to have the number of copies
shown that need to be printed.

The heading in section 1 shows it, so the heading in section 4 also got
given the same field with the same name. This way both fields, as they
both have the same name, get populated at the same time.