Upgraded Rails to 4.2.1
The example from the controller, which was moved into a request is
because previously, with Rails 4.2.0, the headers were being set as per
ActionDispatch setting in config/application.rb:

config.action_dispatch.default_headers

But when testing them in the controller, with Rails 4.2.1, none of those
headers were being set.

In the request spec, the headers *are* being set, hence the move of the
example.